Adds Hash trait and Hasher struct to std lib.
Implements Hash trait for builtin types and for some std lib types.

This is required to have a proper hashing that does not change in
case the internal memory representation of type changes.

Current hashing functions are hashing the types in memory directly, as this
representation might change in future versions of the VM and compiler we
want to have a more reliable way of getting hashes that won't change in
future versions of the compiler and std lib.

Replaces use of sha256 function with Hasher in storage_map and storage_vec.